Implement counter delays
The product spec calls for the effects of passing a counter target to
optionally be delayed by a certain amount of time, for further confidence
that the user will not be reverted before unlocking microcontribution
functionality.  For the initial launch this delay is 48 hours, or 172,800
seconds.

This updates the DAO to check the timestamp for targets_passed entries
before reporting a target as passed.

Change-Id: Ie956f00e5fda4c0fdb6c29dc4bd8dfa114f1c5b7