Add maintenance script to populate entity_description_exists table
Adds a maintenance script to populate the entity_description_exists table
with data on the existence or nonexistence of descriptions in each
langauge for which a Wikidata entity has a linked Wikipedia article.

The script iterates through ips_row_id values in the wb_items_per_site
table in batches in order to limit the sizes of reads and writes and
prevent DB replication lag.

Change-Id: Ia8f13250dc995510c6ce82a0f8346ad3b2d4aafd