Fix: Use READ_LOCKING when evaluating whether to update targets_passed
After a successful qualifying edit, the count for the relevant language
is incremented. Immediately afterward, a follow-up query is performed to
get the total count for all languages, in order to evaluate whether the
user has passed any targets.

Here, the DAO method performing this follow-up query is updated to take
a $flags parameter, and the IDBAccessObject::READ_LOCKING FLAG is passed
in to read from master and lock out writes in order to avoid reading an
invalid value.

Bug: T219084
Change-Id: Iefa37199798091070a15ad017e30b870c14f5d28