Fix revert counting for non-language-specific counters, take 2
The change in d0fd394 did not update revert counting according to the
new treatment for non-language-specific counters; hence, reverts for
those counters (including the new Depicts counter) will not be counted
correctly without this change.

We also need to check that the action in the edit summary matches the
specified action for the counter, or else reverts will be counted once by
each language-agnostic counter.

This patch also adds some new unit tests around edit and revert counting
to increase confidence in the counters and help prevent regressions.
Change tag checking is moved out of the conditionallyIncrementRevertCount
method to facilitate testing. The check for defined( MW_API ) is also
removed; it is redundant with the 'action' parameter check and also
complicates testing.

---

CHANGES FROM 0af18bf: Move change tag addition back into the
conditionallyIncrementEditCount method. Without the conditional checks,
we tag every revision on the wiki (T247479).

Bug: T244974
Change-Id: I5c5b47655100c25a29583637b8fcc5a89a30e11b