Update logic for recording Image Tag edits.
The app now uses the `wbeditentity` action for setting multiple depicts
statements at once. This patch updates the logic to look for that
action, while still checking for the old action (wbsetclaim) since there
might be older versions of the app still using it.

This also simplifies and condenses the general logic of validating edits
for inclusion in the edit counts.

Bug: T258111
Change-Id: I6a5fb8a0692896bfb72f1a4f4e5acc8519d0a0da